backend/requirements.txt
# Core LLM and RAG
langchain==0.1.0
langchain-openai==0.0.2
openai==1.10.0
httpx<0.28.0

# Vector Database - PINECONE
pinecone-client==3.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3

# XML/RSS Parsing
feedparser==6.0.10
lxml==5.1.0

# JSON Schema Validation
jsonschema==4.20.0

# Progress bars
tqdm==4.66.1
